+++ src/statistical/test_ks.cpp
@@ -13,7 +13,7 @@ void TestKS<T_INPUT, T_TIME>::run(const MeasuresPool<T_INPUT, T_TIME> &measures)
 }
 
 template <typename T_INPUT, typename T_TIME>
-unsigned int TestKS<T_INPUT, T_TIME>::get_minimal_sample_size(unsigned short req_power) const {
+unsigned long TestKS<T_INPUT, T_TIME>::get_minimal_sample_size(unsigned short req_power) const {
 
 // TODO Add citation
